requirements.txt
Django<1.10

django-allauth
django-autocomplete-light
django-crispy-forms
django-extensions
django-el-pagination
django-filter

celery[redis]

Pillow
vobject
pygments

# database layer
MySQL-python

# workaround python-dateutil issue (python 2.x / python 3)
python-dateutil
